funções da fila separadas em headers.
