+++ cminado.c
@@ -3,55 +3,16 @@
 #include <stdlib.h>
 #include <time.h>
 #include <unistd.h>
+#include "fila.h"
+#define N 10
+#define posvalida(a,b) (a>=0 && a<N && b>=0 && b<N)
+#define MAXBOMBS 2
 
 struct cel_st {
   char valor;
   char status;
 };
 
-typedef struct Item {
-  int l, c;
-} Item;
-struct fila_st {
-  Item *v;
-  int size;
-  int inicio;
-  int fim;
-  int ocupados;
-};
-
-int inicializafila(struct fila_st *f,int s) {
-  f->v=malloc(sizeof(Item)*s);
-  f->size=s;
-  f->inicio=-1;
-  f->fim=0;
-  f->ocupados=0;
-  return 1;
-}
-
-int enfila(struct fila_st *f,Item e) {
-  if(f->ocupados>=f->size) return 0;
-  f->v[f->fim]=e;
-  f->fim=(f->fim+1)%f->size;
-  f->ocupados++;
-  return 1;
-}
-
-void desenfila(struct fila_st *f) {
-  f->ocupados--;
-  f->inicio= (f->inicio+1)%f->size;
-}
-
-Item espia(struct fila_st *f) {
-  return f->v[(f->inicio+1)%f->size];
-}
-
-int esta_vazia(struct fila_st *f) {
-  return f->ocupados==0;
-}
-
-#define N 10
-#define posvalida(a,b) (a>=0 && a<N && b>=0 && b<N)
 struct cel_st campo[N][N];
 
 void inicializacampo() {
@@ -74,6 +35,7 @@ void imprimecampo() {
   }
   usleep(500000);
 }
+
 void imprimecampoCHEAT() {
   for(int i=0;i<N;i++)
   {
@@ -83,7 +45,6 @@ void imprimecampoCHEAT() {
   }
 }
 
-#define MAXBOMBS 2
 void firstclick(int l,int c) {
   campo[l][c].status=1;
   srand(time(NULL));
@@ -105,11 +66,6 @@ void firstclick(int l,int c) {
         campo[i][j].valor++;
       }
   }
-}
-
-void liberafila(struct fila_st *f) {
-  free(f->v);
-  f->size=0;
 }
 
 void abreclick(int l,int c) {
@@ -158,7 +114,5 @@ int main(void)
     if(!posvalida(l,c)) { printf("COLOQUE LUGAR CERTO DATEBAYO\n");exit(1);}
     abreclick(l,c);
   }
-  //printf("\n\n");
-  //mprimecampoCHEAT();
   return 0;
 }
